Allow Monocle to be loaded through SPL
This is accomplished by making the inner mod a proper JiJ, which
means FML will still find it even if our custom locator is not run
(which is the case when SPL downloads the outer jar).

When run in this configuration, a dependency override must explicitly
be added to `config/fml.toml` in order to disable the Iris DRM, as follows:

```
dependencyOverrides.iris = ["-embeddium"]
```

If we are being loaded directly from the mods folder, the fact that
we are a lowcode mod is completely ignored, as we provide FML service files
in META-INF/services, so the old locator continues to work normally.

Credit to ferriarnus for coming up with the idea for this trick.